requirements.txt
fastapi 
uvicorn 
opencv-python 
pyaudio
google-cloud-storage